Add hikey-optee64 machine
This machine builds an image that runs on the Hikey 96board.
Currently, not all of the necessary parts are built to make a full
image, and it will be necessary to manually install the kernel, dtb,
and ramdisk into a boot image.  However, the trusted firmware bl1.bin,
and fip.bin work with the resulting kernel/initrd, and xtest passes.