Add 'arm-grub' target
The existing OE recipes don't have support for building Grub on
non-x86.  This recipe builds a cross-targeted aarch64 grub.  The
result of the build is a single deployed file 'grubaa64.efi' which is
what the firmware will then boot into.

The recipe is a bit complex because this target is only partially
cross compiled.  The image generation tools are native to the host,
but the target images are built for aarch64.  The normal OE
native.class doesn't do the right thing, so we just re-assign a bunch
of environment variables to get it to properly build.